DSMPropertyBag.h: change integers to pointer size integers, as CAtlArray outputs them that way, and there's pointer logic done with these.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4731 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/DSMSplitter/DSMSplitterFile.cpp
@@ -286,7 +286,7 @@ __int64 CDSMSplitterFile::Read(__int64 len, CStringW& str)
 __int64 CDSMSplitterFile::FindSyncPoint(REFERENCE_TIME rt)
 {
 	if (/*!m_sps.IsEmpty()*/ m_sps.GetCount() > 1) {
-		int i = range_bsearch(m_sps, m_rtFirst + rt);
+		ptrdiff_t i = range_bsearch(m_sps, m_rtFirst + rt);
 		return i >= 0 ? m_sps[i].fp : 0;
 	}
